requirements.2014_1219.txt
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.6.1
Pygments==2.0.1
Sphinx==1.2.3
biopython==1.64
docutils==0.12
multi-key-dict==2.0.1
numpy==1.9.1
pandas==0.15.1
patsy==0.3.0
pdfrw==0.1
progressbar==2.3
psutil==2.1.3
pytabix==0.1
python-dateutil==2.2
pytz==2014.9
reportlab==3.1.8
rpy2==2.5.2
rst2pdf==0.93.dev
scipy==0.14.0
singledispatch==3.4.0.3
six==1.8.0
sphinx-argparse==0.1.13
statsmodels==0.6.0
-e git+git@github.com:koesterr/uga.git@0971e32627be299666f4178bf296eeb986814e5a#egg=uga-master
virtualenv==1.10.1
wsgiref==0.1.2
